requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
loguru
openai
asyncio
pydantic
prometheus-client
pytest

google-generativeai
python-dotenv
colorama
fastapi
uvicorn
websockets
jinja2